Create README.md
File-Box is a website where you can store files and folders you don't need right now. If you have an old folder or file that you can't find on your computer, you can download it again from the website. It’s currently in beta and will be updated with improvements to make it more efficient and user-friendly.